+++ lib/src/web/native_peer_connection_factory.dart
@@ -48,6 +48,9 @@ class NativePeerConnectionFactory {
   Future<void> startLocalRecording() async => throw UnimplementedError();
   Future<void> stopLocalRecording() async => throw UnimplementedError();
 
+  Future<void> setMicrophoneMuted(bool muted) async =>
+      throw UnimplementedError();
+
   Future<void> suspendAudio() async {}
   Future<void> resumeAudio() async {}